Blog/requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==3.2.15
django-autoslug==1.9.8
django-tinymce==3.5.0
Django-Verify-Email==2.0.3
filelock==3.9.0
importlib-metadata==5.2.0
mysql-client==0.0.1
mysqlclient @ file:///D:/mysqlclient-1.4.6-cp37-cp37m-win32.whl
platformdirs==2.6.2
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.4.0
virtualenv==20.17.1
zipp==3.11.0
